Fixed `Tensor` value typing (#14876)
* Fixed types of Tensor.FromArray and Tensor.FromArrayToRef

* Fixed types

* Fixed ThinEngine.setMatrices type
Updated Tensor types

* Removed uneeded DeepImmutable in FromArray and FromArrayToRef

* Added new TensorValue
Removed TensorValueElement and TensorElement
Changed FlattenArray to handle recursive arrays
Renamed TensorValue to ValueOfTensor

* Added a newline

* Update packages/dev/core/src/Engines/Native/nativeDataStream.ts

* Update packages/dev/core/src/Engines/Native/nativeDataStream.ts

---------

+++ packages/dev/core/src/Engines/Native/nativePipelineContext.ts
@@ -488,7 +488,7 @@ export class NativePipelineContext implements IPipelineContext {
      */
     public setMatrix(uniformName: string, matrix: IMatrixLike): void {
         if (this._cacheMatrix(uniformName, matrix)) {
-            if (!this._engine.setMatrices(this._uniforms[uniformName]!, matrix.asArray() as Float32Array)) {
+            if (!this._engine.setMatrices(this._uniforms[uniformName]!, matrix.asArray())) {
                 this._valueCache[uniformName] = null;
             }
         }
